+++ CHANGELOG.md
@@ -11,6 +11,7 @@ Version 0.3.00 - Not released yet
 * Remove wrapper methods from HelperString redirecting to HelperSanitize 
 * Remove wrapper methods from HelperString redirecting to HelperXml 
 * Add method: HelperArray::isAnyKeySet()
+* Update method: HelperZip::zipFiles() set default value of argument $deleteFilesWhenDone to false
 
 Version 0.2.21 - Released 2020/09/28
 ------------------------------------